LayoutTests/platform/gtk/imported/w3c/web-platform-tests/xhr/event-error-order.sub-expected.txt

CONSOLE MESSAGE: Error resolving “nonexistent.web-platform.test”: Name or service not known
CONSOLE MESSAGE: XMLHttpRequest cannot load http://nonexistent.web-platform.test:8800/ due to access control checks.

PASS XMLHttpRequest: event - error (order of events)
